Moved AssetService, AssetResponse, ResourceJsonAssetService and ResourceJsonAssetServiceTest to a different folder, so they are not tied to Sep24.
