+++ core/src/main/java/org/stellar/anchor/asset/ResourceJsonAssetService.java
@@ -1,13 +1,11 @@
-package org.stellar.anchor.plugins.asset;
+package org.stellar.anchor.asset;
 
 import com.google.gson.Gson;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Data;
-import org.stellar.anchor.dto.sep24.AssetResponse;
 import org.stellar.anchor.exception.SepNotFoundException;
-import org.stellar.anchor.sep24.AssetService;
 import org.stellar.anchor.util.FileUtil;
 import org.stellar.anchor.util.Log;
