+++ core/src/main/java/org/stellar/anchor/sep38/Sep38Service.java
@@ -1,7 +1,7 @@
 package org.stellar.anchor.sep38;
 
+import org.stellar.anchor.asset.AssetService;
 import org.stellar.anchor.config.Sep38Config;
-import org.stellar.anchor.sep24.AssetService;
 import org.stellar.anchor.util.Log;
 
 public class Sep38Service {
