+++ core/src/test/kotlin/org/stellar/anchor/asset/ResourceJsonAssetServiceTest.kt
@@ -1,4 +1,4 @@
-package org.stellar.anchor.plugins.asset
+package org.stellar.anchor.asset
 
 import com.google.gson.JsonSyntaxException
 import org.junit.jupiter.api.Assertions.*
