+++ core/src/test/kotlin/org/stellar/anchor/sep24/Sep24ServiceTest.kt
@@ -19,6 +19,8 @@ import org.stellar.anchor.Constants.Companion.TEST_ASSET_ISSUER_ACCOUNT_ID
 import org.stellar.anchor.Constants.Companion.TEST_CLIENT_DOMAIN
 import org.stellar.anchor.Constants.Companion.TEST_TRANSACTION_ID_0
 import org.stellar.anchor.Constants.Companion.TEST_TRANSACTION_ID_1
+import org.stellar.anchor.asset.AssetService
+import org.stellar.anchor.asset.ResourceJsonAssetService
 import org.stellar.anchor.config.AppConfig
 import org.stellar.anchor.config.Sep24Config
 import org.stellar.anchor.dto.sep24.GetTransactionRequest
@@ -27,7 +29,6 @@ import org.stellar.anchor.exception.SepException
 import org.stellar.anchor.exception.SepNotFoundException
 import org.stellar.anchor.exception.SepValidationException
 import org.stellar.anchor.model.Sep24Transaction
-import org.stellar.anchor.plugins.asset.ResourceJsonAssetService
 import org.stellar.anchor.sep10.JwtService
 import org.stellar.anchor.sep10.JwtToken
 import org.stellar.anchor.util.DateUtil
