+++ platform/src/main/java/org/stellar/anchor/server/SepConfig.java
@@ -5,6 +5,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.stellar.anchor.asset.AssetService;
+import org.stellar.anchor.asset.ResourceJsonAssetService;
 import org.stellar.anchor.config.AppConfig;
 import org.stellar.anchor.config.Sep10Config;
 import org.stellar.anchor.config.Sep1Config;
@@ -12,11 +14,9 @@ import org.stellar.anchor.config.Sep38Config;
 import org.stellar.anchor.exception.SepNotFoundException;
 import org.stellar.anchor.filter.Sep10TokenFilter;
 import org.stellar.anchor.horizon.Horizon;
-import org.stellar.anchor.plugins.asset.ResourceJsonAssetService;
 import org.stellar.anchor.sep1.Sep1Service;
 import org.stellar.anchor.sep10.JwtService;
 import org.stellar.anchor.sep10.Sep10Service;
-import org.stellar.anchor.sep24.AssetService;
 import org.stellar.anchor.sep38.Sep38Service;
 
 /** SEP configurations */
